+++ CHANGES.rst
@@ -3,10 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+[0.0.17] - XXXX-XX-XX
+=====================
+
 Fixed
 *****
 
 - Fixed group deletion button. :pr:`80`
+- Fixed post requests in oidc clients views. :pr:`81`
 
 [0.0.16] - 2022-12-15
 =====================